rusk: Move SpentTransaction encoding to the right module
